PCHR-1768: Add Tests for Contract Duration Calculation
Added tests that verify duration of 15th day of the month to 14th day of next
month is always "1 month" for all months in a two year period, even on a leap
year. Also added a test for special case of duration starting on the 29th of
february of a leap year, which should count number of complete years/months
between dates and add leftover days, so duration between 2016-02-29 and
2017-02-28 is "1 year 1 day".

Also, improved docblock of duration() method.